+++ src/clientes/clientes.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { ClientesService } from './service/clientes.service';
-import { ClientesController } from './controller/clientes.controller';
 import { Contas } from '../contas/entities/conta.entity';
 import { Gerente } from '../funcionarios/entities/gerente.entity';
-import { ClientesRepository } from './repository/clientes.repository';
 import { FuncionariosRepository } from '../funcionarios/repository/funcionario.repository';
+import { ClientesController } from './infra/adapters/inbound/controller/clientes.controller';
+import { ClientesService } from './application/service/clientes.service';
+import { ClientesRepository } from './infra/adapters/outbound/repository/clientes.repository';
 
 @Module({
   controllers: [ClientesController],
@@ -12,4 +12,4 @@ import { FuncionariosRepository } from '../funcionarios/repository/funcionario.r
   exports: [ClientesService],
   imports: [Gerente, Contas],
 })
-export class ClientesModule {}
+export class ClientesModule { }
