+++ src/contas/contas.module.ts
@@ -3,9 +3,9 @@ import { ContasService } from './services/contas.service';
 import { ContasController } from './controller/contas.controller';
 import { TransacoesModule } from '../transacoes/transacoes.module';
 import { TransacoesRepository } from '../transacoes/repository/transacoes.repository';
-import { ContasFactory } from '../factories/contas.factory';
-import { TransacoesFactory } from '../factories/transacoes.factory';
 import { ContasRepository } from './repository/contas.repository';
+import { TransacoesFactory } from '../factories/transacoes/transacoes.factory';
+import { ContasFactory } from '../factories/contas/contas.factory';
 
 @Module({
   imports: [forwardRef(() => TransacoesModule)],
