+++ src/factories/testes/funcionarios.factory.spec.ts
@@ -1,4 +1,4 @@
-import { FuncionariosFactory } from '../funcionarios.factory';
+import { FuncionariosFactory } from '../funcionarios/funcionarios.factory';
 import { CreateFuncionarioDto } from '../../funcionarios/dto/create-funcionario.dto';
 import { TipoCargo } from '../../common/enums/tipo-.banco.enum';
 import { Agente } from '../../funcionarios/entities/agente.entity';
