+++ src/funcionarios/dto/create-funcionario.dto.ts
@@ -1,5 +1,5 @@
 import { TipoCargo } from '../../common/enums/tipo-.banco.enum';
-import { Cliente } from '../../clientes/entities/cliente.entity';
+import { Cliente } from '../../clientes/domain/entities/cliente.entity';
 
 export class CreateFuncionarioDto {
   cargo: TipoCargo;
