+++ src/funcionarios/entities/gerente.entity.ts
@@ -1,6 +1,6 @@
 import { Funcionario } from './funcionario.entity';
-import { Cliente } from '../../clientes/entities/cliente.entity';
 import { TipoCargo } from '../../common/enums/tipo-.banco.enum';
+import { Cliente } from '../../clientes/domain/entities/cliente.entity';
 
 export class Gerente extends Funcionario {
   clientes: Cliente[];
