+++ src/transacoes/service/transacoes.service.ts
@@ -5,7 +5,7 @@ import {
 import { Transacao } from '../entities/transacao.entity';
 import { TransacoesRepository } from '../repository/transacoes.repository';
 import { CriarTransactionDto } from '../dto/create-transaction.dto';
-import { TransacoesFactory } from '../../factories/transacoes.factory';
+import { TransacoesFactory } from '../../factories/transacoes/transacoes.factory';
 
 @Injectable()
 export class TransacoesService {
