+++ src/transacoes/transacoes.module.ts
@@ -1,8 +1,8 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TransacoesController } from './controller/transacoes.controller';
 import { TransacoesService } from './service/transacoes.service';
-import { ContasModule } from 'src/contas/contas.module';
-import { TransacoesFactory } from '../factories/transacoes.factory';
+import { ContasModule } from '../contas/contas.module';
+import { TransacoesFactory } from '../factories/transacoes/transacoes.factory';
 import { TransacoesRepository } from './repository/transacoes.repository';
 
 @Module({
